[FIRRTL] Add printf widths, reject unknown (#8378)
Add width specifiers to FIRRTL printfs.  This will allow for width
specifiers on printf format strings which take a radix (binary, decimal,
or hexadecimal).  Additionally, enshrine that `%%` is a literal `%` and
will be treated as such.

`firtool` will now reject unknown width specifiers.  E.g., this closes the
`%m` loophole.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>