[FIRRTL] Parse GC View Annos w/o Circuit
Change the parsing of Grand Central View annotations to _not_ require a
circuit to be included in the "ref" part of the annotation.  Upstream
Chisel has dropped support for the circuit part of targets [[1]] and the
internal Grand Central code base must follow suit.

This change is necessary to make this work with the internal code base.
Note: this permissively ignore a circuit if present, so this is backwards
compatible.

While unnecessary, clean up all tests and documentation to _not_ include
the circuit in the "ref".

[1]: https://github.com/chipsalliance/chisel/commit/4cc4ab81e8d2819a076aed1d0c427fe5bebafb14

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>